model/requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.1.0
python-dotenv>=0.19.0

# SDN/Network
ryu>=4.34
mininet>=2.3.0
scapy>=2.4.5
networkx>=2.6.3

# API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Data Processing
matplotlib>=3.4.3
seaborn>=0.11.2

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
